rename class files; WIP somd restraints
